Restructure & imports optimization
