+++ db_models/sticker_sets.py
@@ -1,7 +1,7 @@
 import sqlalchemy
 from sqlalchemy.orm import relationship
 
-from models.db_session import SqlAlchemyBase
+from db_models.db_session import SqlAlchemyBase
 
 
 class StickerSet(SqlAlchemyBase):
