+++ db_models/stickers.py
@@ -1,7 +1,7 @@
 import sqlalchemy
 from sqlalchemy.orm import relationship
 
-from models.db_session import SqlAlchemyBase
+from db_models.db_session import SqlAlchemyBase
 
 
 class Sticker(SqlAlchemyBase):
