--- models/__all_models.py
@@ -1,5 +1,0 @@
-from models.search_data import SearchData
-from models.set_links import SetLink
-from models.sticker_sets import StickerSet
-from models.stickers import Sticker
-from models.user_sets import UserSet
